requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.30
openai==2.29.0
python-multipart==0.0.9
pillow==10.3.0
python-dotenv==1.0.1
pydantic==2.7.1
aiofiles==23.2.1
httpx==0.27.0
pdf2image==1.17.0
pypdfium2==5.6.0
pytesseract==0.3.10
requests==2.31.0
botframework-connector==4.15.0
botbuilder-core==4.15.0
botbuilder-schema==4.15.0
